+++ apps/example/@types/code-app-env.d.ts
@@ -1,0 +1,8 @@
+import '@brandingbrand/code-app-env';
+
+declare module '@brandingbrand/code-app-env' {
+  interface AppEnvironment {
+    id: string;
+    domain: string;
+  }
+}