+++ apps/example/coderc/env/env.store.ts
@@ -1,0 +1,7 @@
+import type {AppEnvironment} from '@brandingbrand/code-app-env';
+import {defineEnv} from '@brandingbrand/code-cli-kit';
+
+export default defineEnv<AppEnvironment>({
+  id: 'lmn34567',
+  domain: 'https://storeDomain.com',
+});