+++ packages/templates/plugin/src/index.ts
@@ -1,30 +1,37 @@
-/**
- * Defines a plugin for @brandingbrand/code-cli-kit.
- * @module Plugin
- */
-
-import {type BuildConfig, definePlugin} from '@brandingbrand/code-cli-kit';
+import {definePlugin} from '@brandingbrand/code-cli-kit';
 
 /**
- * Defines a plugin with functions for both iOS and Android platforms.
- * @alias module:Plugin
- * @param {Object} build - The build configuration object.
- * @param {Object} options - The options object.
+ * Defines a Flagship™ Code `prebuild` plugin.
+ *
+ * Code plugins are intended to encapsulate native code transformations,
+ * but may also be used to peform any setup task your project requires.
+ *
+ * Plugin functions are always executed sequentially in the following order:
+ * 1. `common` - platform-agnostic code
+ * 2. `ios` - iOS-specific code
+ * 3. `android` - Android-specific code
+ *
+ * Platform functions may not always be executed, depending on the prebuild `--platform` argument.
+ *
+ * Avoid executing platform-specific tasks outside of their respective platform functions.
+ * Failure to maintain this separation may lead to build failures or unexpected behavior.
+ *
+ * Each function receives the same parameters:
+ * - `build`: The current build configuration object.
+ * - `options`: The CLI options provided to the prebuild command.
+ * - `codeConfig`: The project's 'flagship-code.config.ts' file.
  */
 export default definePlugin({
   /**
-   * Function to be executed for iOS platform.
-   * @param {Object} build - The build configuration object for iOS.
-   * @param {Object} options - The options object for iOS.
-   * @returns {Promise<void>} A promise that resolves when the process completes.
+   * Common function to be executed for all platforms
    */
-  ios: async function (build: BuildConfig, options: object): Promise<void> {},
-
+  common: async function (build, options): Promise<void> {},
+  /**
+   * Function to be executed for iOS platform.
+   */
+  ios: async function (build, options): Promise<void> {},
   /**
    * Function to be executed for Android platform.
-   * @param {Object} build - The build configuration object for Android.
-   * @param {Object} options - The options object for Android.
-   * @returns {Promise<void>} A promise that resolves when the process completes.
    */
-  android: async function (build: object, options: object): Promise<void> {},
+  android: async function (build, options): Promise<void> {},
 });